Clarify what the "rest" is contrasting in Google eng-practices documentation.
This is a frequently quoted section for me to during code reviews, and starting
the quote with "The rest of the description..." begs the question of what the
"rest" is contrasted against (the firstline section).  Here we make it explicit
that the "rest" is in contrast to the firstline. This should enhance clarity
when only quoting section in isolation.

PiperOrigin-RevId: 383497526