Update "Think for Yourself" to "Think Collaboratively".
This both updates the header text to better reflect the spirit of the section and expands the section text to emphasize that code review should be a collaborative exercise, not a combative one.

This also mirrors the reviewer guidance about courtesy (https://google.github.io/eng-practices/review/reviewer/comments.html#courtesy) in the author guidance.

PiperOrigin-RevId: 438626072